Extract SelectOptions type to avoid circular imports
