+++ packages/pxweb2-ui/src/lib/components/Select/Select.spec.tsx
@@ -2,7 +2,8 @@ import { render, fireEvent, screen } from '@testing-library/react';
 import { describe, it, expect, beforeEach, vi } from 'vitest';
 import '@testing-library/jest-dom/vitest';
 
-import Select, { SelectOption } from './Select';
+import Select from './Select';
+import { SelectOption } from './SelectOptionType';
 import { mockHTMLDialogElement } from '../../util/test-utils';
 
 describe('Select', () => {
