+++ packages/pxweb2-ui/src/lib/components/Select/SelectOptionType.ts
@@ -1,0 +1,4 @@
+export type SelectOption = {
+  label: string;
+  value: string;
+};
