+++ packages/pxweb2-ui/src/lib/components/VariableBox/utils.ts
@@ -1,4 +1,4 @@
-import { SelectOption } from '../Select/Select';
+import { SelectOption } from '../Select/SelectOptionType';
 import { CodeList } from '../../shared-types/codelist';
 import { mapCodeListsToSelectOptions } from '../../util/util';
