+++ packages/pxweb2-ui/src/lib/components/VariableList/VariableList.tsx
@@ -3,7 +3,7 @@ import cl from 'clsx';
 import styles from './VariableList.module.scss';
 import { SelectedVBValues, VariableBox } from '../VariableBox/VariableBox';
 import { PxTableMetadata } from '../../shared-types/pxTableMetadata';
-import { SelectOption } from '../Select/Select';
+import { SelectOption } from '../Select/SelectOptionType';
 import { Value } from '../../shared-types/value';
 
 export type VariableListProps = {
